% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laxyapi.R
\docType{package}
\name{laxyapi}
\alias{laxyapi}
\title{laxyapi: Interact with the Laxy REST API}
\description{
The laxyapi package simplifies interaction with the Laxy REST API.
It allows downloading of job files and associated metadata.
}
\section{laxyapi functions}{


get_job_files
}

